% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explor.R, R/explor_text_corpus.R,
%   R/explor_text_dfm.R
\name{explor}
\alias{explor}
\alias{explor.Corpus}
\alias{explor.corpus}
\alias{explor.dfm}
\title{Interface for analysis results exploration}
\usage{
explor(obj, ...)

\method{explor}{Corpus}(obj, raw_corpus = NULL, stopwords = NULL,
  thesaurus = NULL, ...)

\method{explor}{corpus}(obj, raw_corpus = NULL, stopwords = NULL,
  thesaurus = NULL, ...)

\method{explor}{dfm}(obj, ...)
}
\arguments{
\item{obj}{corpus object}

\item{...}{arguments passed to other methods}

\item{raw_corpus}{optional raw documents corpus}

\item{stopwords}{stopwords character vector}

\item{thesaurus}{quanteda thesaurus list}
}
\value{
The function launches a shiny app in the system web browser.
}
\description{
This function launches a shiny app in a web browser in order to do
interactive visualisation and exploration of a textual corpus.
}
\examples{
\dontrun{
require(quanteda)
data("data_corpus_inaugural")
explor(data_corpus_inaugural)
}
}
\author{
Julien Barnier <julien.barnier@ens-lyon.fr>
}
