% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_results.R, R/prepare_results_corpus.R
\name{prepare_results}
\alias{prepare_results}
\alias{prepare_results.Corpus}
\alias{prepare_results.corpus}
\title{Analysis results preparation}
\usage{
prepare_results(obj)

\method{prepare_results}{Corpus}(obj)

\method{prepare_results}{corpus}(obj)
}
\arguments{
\item{obj}{object containing analysis results}
}
\description{
This function prepares results to be used by \code{explor}. Not to be used directly.
}
\seealso{
\code{\link[tm]{Corpus}}

\code{\link[quanteda]{corpus}}
}
