% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texplor.R, R/texplor_corpus.R, R/texplor_dfm.R
\name{texplor}
\alias{texplor}
\alias{texplor.Corpus}
\alias{texplor.corpus}
\alias{texplor.dfm}
\title{Interface for analysis results exploration}
\usage{
texplor(obj, ...)

\method{texplor}{Corpus}(obj, raw_corpus = NULL, stopwords = NULL,
  dictionary = NULL, ...)

\method{texplor}{corpus}(obj, raw_corpus = NULL, stopwords = NULL,
  dictionary = NULL, ...)

\method{texplor}{dfm}(obj, ...)
}
\arguments{
\item{obj}{corpus object}

\item{...}{arguments passed to other methods}

\item{raw_corpus}{optional raw documents corpus}

\item{stopwords}{stopwords character vector}

\item{dictionary}{quanteda dictionary list}
}
\value{
The function launches a shiny app in the system web browser.
}
\description{
This function launches a shiny app in a web browser in order to do
interactive visualisation and exploration of a textual corpus.
}
\examples{
\dontrun{
require(quanteda)
data("data_corpus_inaugural")
texplor(data_corpus_inaugural)
}
}
\author{
Julien Barnier <julien.barnier@ens-lyon.fr>
}
