% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{texplor_css}
\alias{texplor_css}
\title{Interface for analysis results exploration}
\usage{
texplor_css()
}
\arguments{
\item{obj}{corpus object}

\item{...}{arguments passed to other methods}
}
\value{
The function launches a shiny app in the system web browser.
}
\description{
This function launches a shiny app in a web browser in order to do
interactive visualisation and exploration of a textual corpus.
}
\examples{
\dontrun{
require(quanteda)
data("data_corpus_inaugural")
texplor(data_corpus_inaugural)
}
}
